+++ R/PredictionDataSurv.R
@@ -6,21 +6,12 @@ as_prediction.PredictionDataSurv = function(x, check = TRUE, ...) { # nolint
 
 #' @export
 check_prediction_data.PredictionDataSurv = function(pdata) { # nolint
-  row_ids = assert_row_ids(pdata$row_id)
-  n = length(row_ids)
-  assert_class(pdata$truth, "Surv")
-
+  n = length(assert_row_ids(pdata$row_id))
+  assert_surv(pdata$truth, "Surv", len = n, any.missing = FALSE, null.ok = TRUE)
+  assert_numeric(pdata$crank, len = n, any.missing = FALSE, null.ok = FALSE)
   assert_numeric(pdata$response, len = n, any.missing = FALSE, null.ok = TRUE)
-  assert_numeric(pdata$crank, len = n, any.missing = FALSE, null.ok = TRUE)
   assert_numeric(pdata$lp, len = n, any.missing = FALSE, null.ok = TRUE)
-
-  if (!is.null(pdata$distr)) {
-    assert_class(pdata$distr, "VectorDistribution")
-    if (is.null(pdata$crank)) {
-      pdata$crank = unname(pdata$distr$mean())
-    }
-  }
-
+  assert_matrix(pdata$distr, nrows = n, any.missing = FALSE, null.ok = TRUE)
   pdata
 }
 
@@ -72,7 +63,7 @@ c.PredictionDataSurv = function(..., keep_duplicates = TRUE) {
   }
 
   if ("distr" %in% predict_types) {
-    result$distr = do.call(c, map(dots, "distr"))
+    result$distr = do.call(rbind, map(dots, "distr"))
   }
 
   set_class(result, "PredictionDataSurv")